requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
alluka==0.1.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
frozenlist==1.4.1
hikari==2.0.0.dev122
hikari-lightbulb==2.3.5
hikari-miru==4.0.0
idna==3.6
multidict==6.0.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pip-review==1.3.0
platformdirs==4.2.0
python-dotenv==1.0.1
soupsieve==2.5
uvloop==0.19.0
yarl==1.9.4
